Remove @supabase/supabase-js and related packages
supabase and mongoDB was clashing had to remove in this branch
Deleted the @supabase/supabase-js package and its dependencies, along with related files from node_modules. Updated other Supabase packages and removed unused dependencies such as array-flatten and bcryptjs. This change likely reflects a migration away from the supabase-js client or a cleanup of unused packages.